Refresh toml module
This required changing from toml4j to jackson-dataformat-toml so this will break configurations which used
custom TomlWriter from toml4j.
Why the library was changed? Well, toml4j was outdated - it was supporting toml version 0.4.0, when we have
1.0.0 . jackson-dataformat-toml supports toml 1.0.0 and does something very important for a config parser
such as ours - retain order. Of course, changing the library had its drawbacks - i've explained in toml/README.md

This change forced another change - that is, force to primitives the values from the map (DefaultSerializer)
when we are deserializing to a object.